refactored load into load_manager service.
