+++ services/monitorer/monitorer_rabbitmq/app/scale.json
@@ -2,7 +2,7 @@
   "min": 1,
   "max": 1,
   "dependences": {
-    "required": [
+    "require": [
 
     ],
     "optional": [
