+++ services/monitorer/trigger/app/scale.json
@@ -2,7 +2,7 @@
   "min": 1,
   "max": 1,
   "dependences": {
-    "required": [
+    "require": [
       {
         "name": "overseer_mongodb",
         "type": "docker",
