+++ services/overseer/dependency_solver/app/scale.json
@@ -2,7 +2,7 @@
   "min": 1,
   "max": 10,
   "dependences": {
-    "required": [
+    "require": [
     ],
     "provide": {
       "dependency_solver:rpc:solve_dependencies": 1,
