+++ services/overseer/load_manager/Makefile
@@ -1,0 +1,4 @@
+SERVICE_NAME ?= $(shell basename $(CURDIR))
+IMAGE_VERSION ?= $(shell cat version).$(BUILD_NUMBER)
+
+include ../../Makefile.inc
