+++ services/overseer/load_manager/app/service/load_manager/tests/test_load_manager.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+import logging
+
+import mock
+# other MS
+import pytest
+
+
+logger = logging.getLogger(__name__)
+
+
+def test_load_manager_method():
+
+    assert 1 + 1 == 2
