+++ services/overseer/load_manager/services.json
@@ -1,0 +1,9 @@
+{
+  "services": [
+    {
+      "name": "{dirname}",
+      "extends": "docker",
+      "type": "[S]overseer"
+    }
+  ]
+}
